requirements.txt
tensorflow==2.3.0
keras==2.4.3
keras_resnet==0.2.0
pillow==8.1.0
numpy==1.18.5
ffmpeg-python==0.2.0
python-dotenv==0.15.0
tqdm==4.56.2
tensorflow-hub==0.7.0
scikit-learn==0.24.1
tensorflow-addons==0.12.1
imgaug==0.4.0
imagecorruptions==1.1.2
adabelief-tf==0.2.0